# src/main/resources/swagger/api.yml
openapi: 3.0.2
info:
  title: Booklib
  version: 1.0.0
paths:
  /book:
    get:
      tags:
        - Book
      responses:
        '200':
          $ref: '#/components/responses/RespGetBookArray'
      operationId: getBooks
      summary: Get books info
    post:
      requestBody:
        content:
          multipart/form-data:
            schema:
              required:
                - file
              type: object
              properties:
                file:
                  format: binary
                  type: string
      tags:
        - Book
      parameters:
        - name: isbn
          description: ''
          schema:
            type: string
          in: query
          required: true
      responses:
        '201':
          description: Book uploaded
      operationId: uploadBook
      summary: Upload book
  /page:
    get:
      tags:
        - Page
      parameters:
        - name: isbn
          description: ''
          schema:
            type: string
          in: query
          required: true
        - name: page
          description: ''
          schema:
            type: integer
          in: query
          required: true
      responses:
        '200':
          $ref: '#/components/responses/RespGetPage'
      operationId: getPageUrl
      summary: Get book page expirable URL
  '/url/{hash}':
    get:
      tags:
        - Url
      responses:
        '200':
          content:
            image/jpeg:
              schema:
                format: binary
                type: string
          description: Logo image in PNG format
      operationId: getPageImage
      summary: Get page image
    parameters:
      - name: hash
        schema:
          type: string
        in: path
        required: true
components:
  schemas:
    Book:
      description: ''
      type: object
      properties:
        isbn:
          description: ''
          type: string
        processed:
          description: ''
          type: boolean
        pageCount:
          description: ''
          type: integer
    Page:
      description: ''
      type: object
      properties:
        isbn:
          description: ''
          type: string
        url:
          description: ''
          type: string
        page:
          description: ''
          type: integer
  responses:
    RespGetBookArray:
      content:
        application/json:
          schema:
            type: array
            items:
              $ref: '#/components/schemas/Book'
      description: ''
    RespGetPage:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Page'
      description: ''
  securitySchemes:
    jwt:
      scheme: bearer
      type: http
security:
  - jwt: []
tags:
  - name: Book
    description: ''
  - name: Page
    description: ''
  - name: Url
    description: ''